Add error handling to prevent errors
I'm sorry I don't have a repro, but I have seen hundreds of our user stack traces in Sentry.  TypeErro Cannot read property 'offsetHeight' of null in adaptHeight().  Maybe a rare timing scenario when the grid is initializing can cause this selector to return null or undefined.  Code just defaults to empty object in this case so it doesn't hit a JavaScript error.